New Controller File That Works
Created a compeltely new controller file. This one is very similar to
the old controller, but it uses Serial.Read instead of Serial.ParseInt.
In order to make Serial.Read work a delay of 5 microseconds is required
at the beginning of each case statement to avoid getting the console
spammed with -1.  Serial.Read will only read one byte at a time, thus
the need to do math to get numbers in the right position.